refactor: move test to file
